Docstrings and refactoring of toc_config
